+++ Text/Diff/Engine/shell.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Class used internally by Diff to actually compute the diffs.
  *