+++ Text/Diff/Renderer.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * A class to render Diffs in different formats.
  *