+++ Text/Diff/Renderer/Inline.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * "Inline" diff renderer.
  *